+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Documentation in German ([#5](https://github.com/scm-manager/scm-webhook-plugin/pull/5))
 
+### Changed
+- Use expression language parser from scm-el-plugin ([#10](https://github.com/scm-manager/scm-webhook-plugin/pull/10))
+
 ## [2.0.0] - 2020-06-04
 ### Changed
 - Changeover to MIT license ([#4](https://github.com/scm-manager/scm-webhook-plugin/pull/4))